Fixed signed / unsigned comparisons

+++ tests/set_test.c
@@ -11,7 +11,7 @@ ssize_t values0_19[] = {
     16, 17, 18, 19
 };
 
-#define MAX_LEN sizeof_array(values0_19)
+#define MAX_LEN (int)sizeof_array(values0_19)
 
 ssize_t values0_19_fill[MAX_LEN + 1];
